Source cleanup and documentation
