+++ gui/src/database/database.cpp
@@ -2,245 +2,21 @@
 #include "conversion.h"
 #include "database/components.h"
 #include "database/database_notification.h"
-#include "utilities/math_utility.h"
+#include "database/database_clone.h"
 
-#include "simdata_io.hpp"
 #include "simulation_data_api.hpp"
 
 #include <algorithm>
 #include <optional>
 #include <unordered_map>
+#include <unordered_set>
 
-#define GLM_ENABLE_EXPERIMENTAL 1
-#include <glm/gtx/io.hpp>
-
-#include <QDateTime>
 #include <QDebug>
-
-template <class T>
-void hash_combine(size_t& seed, T const& v) {
-    seed ^= std::hash<T>()(v) + 0x9e3779b9 + (seed << 6) + (seed >> 2);
-}
-
-template <class T>
-std::optional<T> optional_ptr(T* ptr) {
-    if (ptr) return *ptr;
-    return std::nullopt;
-}
-
-// template <>
-// struct std::hash<SD::OpticalProperties> {
-//     std::size_t operator()(SD::OpticalProperties const& a) const {
-//         size_t seed = 0;
-
-//         hash_combine(seed, a.my_type);
-//         hash_combine(seed, a.error_distribution_type);
-//         hash_combine(seed, a.transmissivity);
-//         hash_combine(seed, a.reflectivity);
-//         hash_combine(seed, a.slope_error);
-//         hash_combine(seed, a.specularity_error);
-//         hash_combine(seed, a.refraction_index_front);
-//         hash_combine(seed, a.refraction_index_back);
-
-//         return seed;
-//     }
-// };
-
-// template <>
-// struct std::hash<db::MaterialComponent> {
-//     std::size_t operator()(db::MaterialComponent const& a) const {
-//         size_t seed = 0;
-
-//         hash_combine(seed, a.optics_front);
-//         hash_combine(seed, a.optics_back);
-
-//         return seed;
-//     }
-// };
-
-template <>
-struct std::hash<db::GeometryComponent> {
-    std::size_t operator()(db::GeometryComponent const& a) const {
-        size_t seed = 0;
-
-        if (a.aperture) hash_combine(seed, *a.aperture);
-        if (a.surface) hash_combine(seed, *a.surface);
-
-        return seed;
-    }
-};
-
-static bool is_equal(SD::aperture_ptr const& a, SD::aperture_ptr const& b);
-static bool is_equal(SD::surface_ptr const& a, SD::surface_ptr const& b);
-
-static size_t hash_aperture(SD::Aperture const& a);
-static size_t hash_surface(SD::Surface const& a);
-
-template <>
-struct std::hash<SD::Aperture> {
-    std::size_t operator()(SD::Aperture const& a) const {
-        return hash_aperture(a);
-    }
-};
-
-template <>
-struct std::hash<SD::Surface> {
-    std::size_t operator()(SD::Surface const& a) const {
-        return hash_surface(a);
-    }
-};
-
-// --- Hashing ----------------------------------------------------------------
-
-static size_t hash_aperture(SD::Aperture const& a) {
-    size_t seed = 0;
-    hash_combine(seed, a.my_type);
-
-    switch (a.my_type) {
-    case SolTrace::Data::ANNULUS: {
-        auto* aa = dynamic_cast<SD::Annulus const*>(&a);
-        if (!aa) break;
-        hash_combine(seed, aa->inner_radius);
-        hash_combine(seed, aa->outer_radius);
-        hash_combine(seed, aa->arc_angle);
-        break;
-    }
-    case SolTrace::Data::CIRCLE: {
-        auto* aa = dynamic_cast<SD::Circle const*>(&a);
-        if (!aa) break;
-        hash_combine(seed, aa->diameter);
-        break;
-    }
-    case SolTrace::Data::HEXAGON: {
-        auto* aa = dynamic_cast<SD::Hexagon const*>(&a);
-        if (!aa) break;
-        hash_combine(seed, aa->circumscribe_diameter);
-        break;
-    }
-    case SolTrace::Data::RECTANGLE: {
-        auto* aa = dynamic_cast<SD::Rectangle const*>(&a);
-        if (!aa) break;
-        hash_combine(seed, aa->x_length());
-        hash_combine(seed, aa->y_length());
-        hash_combine(seed, aa->x_coord());
-        hash_combine(seed, aa->y_coord());
-        break;
-    }
-    case SolTrace::Data::EQUILATERAL_TRIANGLE: {
-        auto* aa = dynamic_cast<SD::EquilateralTriangle const*>(&a);
-        if (!aa) break;
-        hash_combine(seed, aa->circumscribe_diameter);
-        break;
-    }
-    case SolTrace::Data::SINGLE_AXIS_CURVATURE_SECTION: break;
-    case SolTrace::Data::IRREGULAR_TRIANGLE: {
-        auto* aa = dynamic_cast<SD::IrregularTriangle const*>(&a);
-        if (!aa) break;
-        hash_combine(seed, aa->x1);
-        hash_combine(seed, aa->y1);
-        hash_combine(seed, aa->x2);
-        hash_combine(seed, aa->y2);
-        hash_combine(seed, aa->x3);
-        hash_combine(seed, aa->y3);
-        break;
-    }
-    case SolTrace::Data::IRREGULAR_QUADRILATERAL: {
-        auto* aa = dynamic_cast<SD::IrregularQuadrilateral const*>(&a);
-        if (!aa) break;
-        hash_combine(seed, aa->x1);
-        hash_combine(seed, aa->y1);
-        hash_combine(seed, aa->x2);
-        hash_combine(seed, aa->y2);
-        hash_combine(seed, aa->x3);
-        hash_combine(seed, aa->y3);
-        hash_combine(seed, aa->x4);
-        hash_combine(seed, aa->y4);
-        break;
-    }
-    case SolTrace::Data::APERTURE_UNKNOWN: break;
-    }
-
-    return seed;
-}
-
-static size_t hash_surface(SD::Surface const& a) {
-    size_t seed = 0;
-    hash_combine(seed, a.my_type);
-
-    switch (a.my_type) {
-    case SolTrace::Data::CONE: {
-        auto* aa = dynamic_cast<SD::Cone const*>(&a);
-        if (!aa) break;
-        hash_combine(seed, aa->half_angle);
-        break;
-    }
-    case SolTrace::Data::CYLINDER: {
-        auto* aa = dynamic_cast<SD::Cylinder const*>(&a);
-        if (!aa) break;
-        hash_combine(seed, aa->radius);
-        break;
-    }
-    case SolTrace::Data::FLAT: break;
-    case SolTrace::Data::PARABOLA: {
-        auto* aa = dynamic_cast<SD::Parabola const*>(&a);
-        if (!aa) break;
-        hash_combine(seed, aa->focal_length_x);
-        hash_combine(seed, aa->focal_length_y);
-        break;
-    }
-    case SolTrace::Data::SPHERE: {
-        auto* aa = dynamic_cast<SD::Sphere const*>(&a);
-        if (!aa) break;
-        hash_combine(seed, aa->vertex_curv);
-        break;
-    }
-    case SolTrace::Data::HYPER:
-    case SolTrace::Data::GENERAL_SPENCER_MURTY:
-    case SolTrace::Data::TORUS:
-    case SolTrace::Data::SURFACE_UNKNOWN: break;
-    }
-
-    return seed;
-}
+#include <QSet>
 
 namespace db {
 
-TransformComponent extract_tf(SD::Element const& e) {
-    auto pos = e.get_origin_ref();
-
-    auto aim = e.get_aim_vector_ref();
-
-    aim = glm::normalize(aim - pos);
-
-    auto quat = dir_roll_to_quat(aim, e.get_zrot_radians());
-
-    // TODO: FIX
-    // I dont understand why the above is incorrect and the below is right
-    auto quat2 = glm::quat_cast(e.get_local_to_reference());
-
-    // qDebug() << quat << quat2;
-
-    return TransformComponent { .position = pos, .rotation = quat2 };
-}
-
-TransformComponent extract_tf_stage(SD::Element const& e) {
-    auto pos = e.get_origin_stage();
-
-    auto aim = e.get_aim_vector_stage();
-
-    aim = glm::normalize(aim - pos);
-
-    auto quat = dir_roll_to_quat(aim, e.get_zrot_radians());
-
-    auto quat2 = glm::quat_cast(e.get_local_to_reference());
-
-    // qDebug() << quat << quat2;
-
-    return TransformComponent { .position = pos, .rotation = quat2 };
-}
-
-// =============================================================================
-
+/// Get a global transform regardless if the entity has a local transform
 static GlobalTransformComponent
 compute_global_without_local_transform(entt::registry& reg,
                                        entt::entity    entity) {
@@ -254,6 +30,7 @@ compute_global_without_local_transform(entt::registry& reg,
     };
 }
 
+/// Compute global transforms for a subtree
 static void update_global_transform_subtree(
     entt::registry&                         reg,
     entt::entity                            entity,
@@ -274,10 +51,12 @@ static void update_global_transform_subtree(
     }
 }
 
+/// On local transform change, update globals
 static void tf_change_callback(entt::registry& reg, entt::entity entity) {
     update_global_transform_subtree(reg, entity);
 }
 
+/// On local transform change, update globals
 static void tf_destroy_callback(entt::registry& reg, entt::entity entity) {
     update_global_transform_subtree(
         reg, entity, compute_global_without_local_transform(reg, entity));
@@ -328,478 +107,9 @@ Database::~Database() {
     qDebug() << Q_FUNC_INFO;
 }
 
-// =============================================================================
-
-struct EntityMapper {
-    entt::registry&                                new_registry;
-    std::unordered_map<entt::entity, entt::entity> old_to_new_map;
-
-    entt::entity operator()(entt::entity e) {
-        auto item = old_to_new_map.find(e);
-
-        if (item == old_to_new_map.end()) {
-            auto new_e        = new_registry.create();
-            old_to_new_map[e] = new_e;
-            return new_e;
-        }
-
-        return item->second;
-    }
-};
-
-template <class T>
-void copy_marker_component(entt::registry const& from,
-                           EntityMapper&         mapper,
-                           entt::registry&       to) {
-    for (auto [e] : from.view<T>().each()) {
-        to.emplace_or_replace<T>(mapper(e));
-    }
-}
-
-template <class T>
-void copy_plain_component(entt::registry const& from,
-                          EntityMapper&         mapper,
-                          entt::registry&       to) {
-    for (auto [e, c] : from.view<T>().each()) {
-        to.emplace_or_replace<T>(mapper(e), c);
-    }
-}
-
-template <class T>
-void copy_nested_component(entt::registry const& from,
-                           EntityMapper&         mapper,
-                           entt::registry&       to) {
-    for (auto [e, c] : from.view<T>().each()) {
-        T component_copy = c;
-        component_copy.remap_entities(mapper);
-        to.emplace_or_replace<T>(mapper(e), component_copy);
-    }
-}
-
-struct DatabaseCloneResult {
-    std::unique_ptr<Database>                        database;
-    std::unordered_map<entt::entity, entt::entity> old_to_new_map;
-};
-
-static DatabaseCloneResult
-clone_database_with_entity_map(Database const& from,
-                               QString         new_database_name,
-                               QObject*        p) {
-    auto  ret           = std::make_unique<Database>(new_database_name, p);
-    auto& from_registry = from.as_registry();
-    auto& to_registry   = ret->as_registry();
-
-    EntityMapper mapper {
-        .new_registry   = to_registry,
-        .old_to_new_map = {},
-    };
-
-    copy_marker_component<InvisibleComponent>(
-        from_registry, mapper, to_registry);
-
-    copy_marker_component<DisabledComponent>(
-        from_registry, mapper, to_registry);
-
-    copy_marker_component<VirtualTagComponent>(
-        from_registry, mapper, to_registry);
-
-    copy_plain_component<IdentityComponent>(
-        from_registry, mapper, to_registry);
-
-
-    copy_marker_component<ElementComponent>(
-        from_registry, mapper, to_registry);
-
-
-    copy_nested_component<ChildOfComponent>(
-        from_registry, mapper, to_registry);
-
-    copy_nested_component<ChildrenComponent>(
-        from_registry, mapper, to_registry);
-
-
-    copy_plain_component<TransformComponent>(
-        from_registry, mapper, to_registry);
-
-    copy_plain_component<GlobalTransformComponent>(
-        from_registry, mapper, to_registry);
-
-
-    ret->ray_source_resource.set(from.ray_source_resource.require().clone());
-    ret->simulation_parameters_resource.set(
-        from.simulation_parameters_resource.require());
-
-
-    copy_plain_component<MaterialComponent>(
-        from_registry, mapper, to_registry);
-
-    copy_nested_component<MaterialGroupComponent>(
-        from_registry, mapper, to_registry);
-
-    copy_nested_component<MaterialGroupMemberComponent>(
-        from_registry, mapper, to_registry);
-
-    // GeometryComponent
-    {
-        for (auto [e, c] : from_registry.view<GeometryComponent>().each()) {
-            auto local = c.clone();
-            to_registry.emplace<GeometryComponent>(mapper(e), local);
-        }
-    }
-
-    copy_nested_component<GeometryGroupComponent>(
-        from_registry, mapper, to_registry);
-
-    copy_nested_component<GeometryGroupMemberComponent>(
-        from_registry, mapper, to_registry);
-
-    copy_marker_component<TagComponent>(
-        from_registry, mapper, to_registry);
-
-    // ATagMemberComponent
-    {
-        // we need to know what all tags are out there
-
-        auto all_tags = QSet<entt::entity>();
-
-        for (auto [e, c] :
-             from_registry.view<TagMembershipComponent>().each()) {
-            for (auto const& t : c.tags) {
-                all_tags.insert(t);
-            }
-        }
-
-        for (auto tag_ent : std::as_const(all_tags)) {
-            auto from_storage = from_registry.storage<ATagMemberComponent>(
-                entt::to_integral(tag_ent));
-
-            auto& to_storage = to_registry.storage<ATagMemberComponent>(
-                entt::to_integral(mapper(tag_ent)));
-
-            for (auto element : *from_storage) {
-                to_storage.emplace(mapper(element));
-            }
-        }
-    }
-
-    copy_nested_component<TagMembershipComponent>(
-        from_registry, mapper, to_registry);
-
-    copy_plain_component<ImportErrorComponent>(
-        from_registry, mapper, to_registry);
-
-    copy_marker_component<SelectedComponent>(
-        from_registry, mapper, to_registry);
-
-    copy_plain_component<ColorComponent>(
-        from_registry, mapper, to_registry);
-
-    copy_plain_component<HasFluxMapComponent>(
-        from_registry, mapper, to_registry);
-
-    return DatabaseCloneResult {
-        .database       = std::move(ret),
-        .old_to_new_map = std::move(mapper.old_to_new_map),
-    };
-}
-
 Database* Database::clone(QString new_database_name, QObject* p) const {
     return clone_database_with_entity_map(*this, new_database_name, p)
         .database.release();
-}
-
-// =============================================================================
-
-struct MaterialMap {
-    std::unordered_map<std::shared_ptr<const SD::OpticalPropertySet>,
-                       entt::entity>
-        map;
-};
-
-static void import_optics(
-    Database&                                            reg,
-    entt::entity                                         entity,
-    SD::Element const&                                   item,
-    MaterialMap&                                         material_groups,
-    std::unordered_map<GeometryComponent, entt::entity>& geometry_groups,
-    size_t&                                              group_counter) {
-
-    auto& registry = reg.as_registry();
-
-    auto property_sptr = item.get_optical_property_set();
-
-    // If it has nothing, dont import
-    if (!property_sptr and !item.get_aperture() and !item.get_surface()) {
-        // does not have geometry.
-        qDebug() << "Skipping optics on" << entt::to_integral(entity);
-        return;
-    }
-
-    // It should have all or nothing
-    if (!property_sptr or !item.get_aperture() or !item.get_surface()) {
-
-        emplace_patch<ImportErrorComponent>(reg, entity, [](auto& c) {
-            c.reason += "Missing aperture and surface";
-        });
-        qWarning() << "Entity" << entt::to_integral(entity)
-                   << "missing surface properties";
-        return;
-    }
-
-    // auto local_mat = MaterialComponent {
-    //     .optics_front = *item.get_front_optical_properties(),
-    //     .optics_back  = *item.get_back_optical_properties(),
-    // };
-
-    auto local_geom = GeometryComponent {
-        .aperture = item.get_aperture(),
-        .surface  = item.get_surface(),
-    };
-
-    if (auto group = material_groups.map.find(property_sptr);
-        group != material_groups.map.end()) {
-        // we have such a group
-
-        reg.assign_material(entity, group->second);
-
-    } else {
-        // no such group
-
-        auto new_group_params = MaterialComponent { *property_sptr };
-
-        auto new_group = MaterialGroupComponent { };
-
-        auto group_entity = reg.create();
-
-        registry.emplace<MaterialGroupComponent>(group_entity, new_group);
-        registry.emplace<MaterialComponent>(group_entity, new_group_params);
-        auto material_name = QString::fromStdString(property_sptr->get_name());
-        if (material_name.isEmpty()) {
-            material_name = QString("Material %1").arg(group_counter);
-        }
-
-        registry.emplace<IdentityComponent>(
-            group_entity, IdentityComponent { .name = material_name });
-
-        group_counter++;
-
-        material_groups.map.try_emplace(property_sptr, group_entity);
-
-        reg.assign_material(entity, group_entity);
-    }
-
-    if (auto group = geometry_groups.find(local_geom);
-        group != geometry_groups.end()) {
-        // we have such a group
-
-        reg.assign_geometry(entity, group->second);
-
-    } else {
-        // no such group
-
-        auto new_group_params = local_geom;
-
-        auto new_group = GeometryGroupComponent { };
-
-        auto group_entity = reg.create();
-
-        registry.emplace<GeometryGroupComponent>(group_entity, new_group);
-        registry.emplace<GeometryComponent>(group_entity, new_group_params);
-        registry.emplace<IdentityComponent>(
-            group_entity,
-            IdentityComponent {
-                .name = QString("Geometry %1").arg(group_counter),
-            });
-
-        group_counter++;
-
-        geometry_groups.try_emplace(local_geom, group_entity);
-
-        reg.assign_geometry(entity, group_entity);
-    }
-}
-
-struct StageComponent {
-    TransformComponent stage_tf;
-    TransformComponent this_in_stage;
-};
-
-void Database::import(SD::SimulationData& data, bool legacy_import) {
-
-    // Assuming we are not re-using registries, which we are not for the moment
-    auto imported_tag = create_tag("imported");
-
-    // we use pointers as element IDs are scoped to global AND composite
-    std::unordered_map<SD::Element*, entt::entity> element_to_entity;
-
-    auto get_or_create_entity = [&](SD::Element* ptr) -> entt::entity {
-        if (!ptr) return entt::null;
-
-        if (auto eiter = element_to_entity.find(ptr);
-            eiter != element_to_entity.end()) {
-            return eiter->second;
-        } else {
-            auto ent               = m_registry.create();
-            element_to_entity[ptr] = ent;
-
-            m_registry.emplace<ElementComponent>(ent);
-
-            assign_tag(ent, imported_tag);
-
-            return ent;
-        }
-    };
-
-    // MAP MAY NOT BE RE-USED!
-    MaterialMap                                         material_groups;
-    std::unordered_map<GeometryComponent, entt::entity> geometry_groups;
-
-    size_t group_counter = 0;
-
-    for (auto iter = data.get_iterator(); !data.is_at_end(iter); ++iter) {
-        auto const& element = *(iter->second);
-
-        // if (element.get_name() == "0") {
-        //     qDebug() << "CHECK 0" << element.is_composite()
-        //              << element.is_single() << element.get_surface().get()
-        //              << element.get_aperture().get();
-        // }
-
-        if (element.is_stage()) {
-            // we want to avoid materializing stage elements
-            auto c = std::dynamic_pointer_cast<SD::StageElement>(iter->second);
-
-            TransformComponent stage_tf = extract_tf(element);
-
-            for (auto iter = c->get_const_iterator(); !c->is_at_end(iter);
-                 ++iter) {
-
-                auto child_ent = get_or_create_entity(iter->second.get());
-
-                m_registry.emplace_or_replace<StageComponent>(
-                    child_ent,
-                    StageComponent {
-                        .stage_tf      = stage_tf,
-                        .this_in_stage = extract_tf_stage(*iter->second),
-                    });
-            }
-
-            // we do NOT add these as children, we instead make them globals
-            continue;
-        }
-
-        entt::entity ent = get_or_create_entity(iter->second.get());
-
-        if (!m_registry.valid(ent)) {
-            qWarning() << "Unable to mirror element" << &element;
-        }
-
-        // auto position = element.get_origin_ref();
-
-        // auto rotation = convert(element.get_local_to_reference());
-
-        m_registry.emplace<TransformComponent>(ent, extract_tf(element));
-
-        if (!element.get_name().empty()) {
-
-            // for legacy, sometimes the name is just a number
-
-            auto name = QString::fromStdString(element.get_name());
-
-            bool ok = false;
-
-            name.toLong(&ok);
-
-            if (legacy_import && ok) {
-                // these are not super useful. imbue with extra, if possible
-
-                auto id = element.get_id();
-
-                name = QStringLiteral("Element %1").arg(id);
-            }
-
-            m_registry.emplace<IdentityComponent>(ent, name);
-        }
-
-        if (!element.is_enabled()) {
-            m_registry.emplace<DisabledComponent>(ent);
-        }
-
-        if (element.is_virtual()) {
-            qDebug() << "Adding virtual tag to " << ent;
-            m_registry.emplace<VirtualTagComponent>(ent);
-        }
-
-        if (element.is_composite()) {
-            auto& c = *static_cast<SD::CompositeElement const*>(&element);
-
-            for (auto iter = c.get_const_iterator(); !c.is_at_end(iter);
-                 ++iter) {
-
-                auto child_ent = get_or_create_entity(iter->second.get());
-
-                if (m_registry.any_of<ChildOfComponent>(child_ent)) {
-                    // Uh oh. We should not have multiple parents!
-                    throw std::runtime_error("Multiple parents for element");
-                }
-
-                set_parent(child_ent, ent);
-            }
-        }
-
-        import_optics(*this,
-                      ent,
-                      element,
-                      material_groups,
-                      geometry_groups,
-                      group_counter);
-    }
-
-    {
-        // burn in globals from a stage
-
-        auto view = m_registry.view<StageComponent const>();
-
-        for (auto const& [e, stage] : view.each()) {
-            // qDebug() << entt::to_integral(e);
-
-            auto new_pos =
-                stage.stage_tf.position +
-                stage.stage_tf.rotation * stage.this_in_stage.position;
-            auto new_rot =
-                stage.stage_tf.rotation * stage.this_in_stage.rotation;
-
-            // qDebug() << new_pos << new_rot;
-            // qDebug() << stage.this_in_stage.position
-            //          << stage.this_in_stage.rotation;
-
-            m_registry.emplace_or_replace<TransformComponent>(
-                e,
-                TransformComponent {
-                    .position = new_pos,
-                    .rotation = new_rot,
-                });
-        }
-
-        m_registry.clear<StageComponent>();
-    }
-
-    ray_source_resource.set(RaySourceResource {
-        .source = data.get_ray_source(),
-    });
-
-    simulation_parameters_resource.set(
-        data.get_simulation_parameters());
-
-    qInfo() << "Imported" << this->m_registry.view<ElementComponent>()->size()
-            << "elements";
-
-    qInfo() << "Imported"
-            << this->m_registry.view<MaterialGroupComponent>()->size()
-            << "materials";
-    qInfo() << "Imported"
-            << this->m_registry.view<GeometryGroupComponent>()->size()
-            << "geometries";
 }
 
 Database::operator entt::registry&() {
@@ -816,263 +126,6 @@ entt::registry& Database::as_registry() {
 
 const entt::registry& Database::as_registry() const {
     return m_registry;
-}
-
-// TODO: remove once we have sorted the vector aim thing
-static void install_transform(SD::element_ptr           ptr,
-                              TransformComponent const& tf_comp) {
-
-    glm::dvec3 aim  = { };
-    double     roll = { };
-    quat_to_dir_roll(tf_comp.rotation, aim, roll);
-
-    auto origin = tf_comp.position;
-    ptr->set_origin(origin.x, origin.y, origin.z);
-
-    aim = (origin + aim * 100.0);
-
-    ptr->set_aim_vector(aim.x, aim.y, aim.z);
-
-    ptr->set_zrot_radians(roll);
-}
-
-static void install_group(SD::element_ptr                 ptr,
-                          SD::OpticalPropertySetReference param,
-                          GeometryComponent const&        geom_param) {
-    ptr->set_aperture(geom_param.aperture);
-    ptr->set_surface(geom_param.surface);
-
-    ptr->set_optical_property_set(param);
-}
-
-static SD::OpticalPropertySet
-with_optical_name(SD::OpticalPropertySet const& optics, QString const& name) {
-    nlohmann::ordered_json node;
-    optics.write_json(node);
-    node["my_name"] = name.toStdString();
-    return SD::OpticalPropertySet(node);
-}
-
-Result<std::shared_ptr<DatabaseExport>, QString> Database::export_to_simdata() {
-    SD::SimulationData ret;
-
-    auto param_ptr = simulation_parameters_resource.get();
-    if (!param_ptr) {
-        return QStringLiteral("Internal error: simulation is missing "
-                              "simulation parameters resource.");
-    }
-
-    ret.get_simulation_parameters() = *param_ptr;
-
-    auto ray_source_ptr = ray_source_resource.get();
-    if (!ray_source_ptr) {
-        return QStringLiteral("Ray source is not defined for this simulation.");
-    }
-
-    ret.add_ray_source(ray_source_ptr->source);
-
-    {
-        auto view = m_registry.view<const ElementComponent>();
-        if (view->size() == 0) {
-            return QStringLiteral("There are no elements in this simulation.");
-        }
-    }
-
-    std::unordered_map<entt::entity, SD::element_ptr> entity_element_map;
-
-    // Mirror all elements
-
-    {
-        auto view = m_registry.view<const ElementComponent>();
-        for (auto const& [e] : view.each()) {
-            SD::element_ptr ptr;
-
-            if (children_of(e).size()) {
-                auto n = std::make_shared<SD::CompositeElement>();
-                ptr    = n;
-            } else if (m_registry.all_of<VirtualTagComponent>(e)) {
-                auto n = std::make_shared<SD::VirtualElement>();
-                ptr    = n;
-            } else {
-                auto n = std::make_shared<SD::SingleElement>();
-                ptr    = n;
-            }
-
-            ptr->set_name(name_of(e).toStdString());
-
-            entity_element_map[e] = ptr;
-        }
-    }
-
-    // qDebug() << Q_FUNC_INFO << entity_element_map.size();
-    // qDebug() << Q_FUNC_INFO << ret.get_number_of_elements();
-
-    {
-        auto view = m_registry.view<const TransformComponent>();
-        for (auto const& [e, tf] : view.each()) {
-            install_transform(entity_element_map.at(e), tf);
-        }
-    }
-
-    {
-        auto view =
-            m_registry.view<const ElementComponent, const IdentityComponent>();
-        for (auto const& [e, ident] : view.each()) {
-            entity_element_map.at(e)->set_name(ident.name.toStdString());
-        }
-    }
-
-    {
-        auto view = m_registry.view<const DisabledComponent>();
-        for (auto const& [e] : view.each()) {
-            entity_element_map.at(e)->disable();
-        }
-    }
-
-    std::unordered_map<entt::entity, SD::OpticalPropertySetReference> prop_map;
-
-    {
-        auto view = m_registry.view<const MaterialGroupComponent,
-                                    const MaterialComponent>();
-        for (auto const& [e, mat_members, mat_data] : view.each()) {
-            auto named_optics =
-                with_optical_name(mat_data.optics, name_of(e));
-            auto ref = ret.find_or_add_optical_property_set(named_optics);
-
-            prop_map.try_emplace(e, ref);
-        }
-    }
-
-    {
-        auto view = m_registry.view<const MaterialGroupMemberComponent,
-                                    const GeometryGroupMemberComponent>();
-        for (auto const& [e, mat, geom] : view.each()) {
-
-            // get group, we assume this is valid
-            auto const& mat_group = prop_map.at(mat.group);
-            auto const& geom_group =
-                m_registry.get<GeometryComponent>(geom.group);
-
-            auto element = entity_element_map.at(e);
-
-            if (auto ptr = m_registry.try_get<ChildrenComponent>(e); ptr) {
-                if (!ptr->children.empty()) {
-                    // it has children AND a group component. add a proxy that
-                    // will only hold the geometry
-
-                    auto composite =
-                        dynamic_cast<SD::CompositeElement*>(element.get());
-                    Q_ASSERT(composite);
-
-                    auto n = std::make_shared<SD::SingleElement>();
-                    composite->add_element(n);
-                    install_group(n, mat_group, geom_group);
-
-                    continue;
-                }
-            }
-
-            // has no children
-
-            // if (element->get_name() == "0") {
-            //     qDebug() << "Install" << entt::to_integral(e)
-            //              << element->get_name() <<
-            //              entt::to_integral(mat.group)
-            //              << entt::to_integral(geom.group);
-            // }
-
-
-            install_group(element, mat_group, geom_group);
-        }
-    }
-
-    {
-        auto view = m_registry.view<const ChildrenComponent>();
-        for (auto const& [e, children] : view.each()) {
-
-            auto composite = dynamic_cast<SD::CompositeElement*>(
-                entity_element_map.at(e).get());
-            Q_ASSERT(composite);
-
-            for (auto child : children.children) {
-
-                auto iter = entity_element_map.find(child);
-
-                if (iter == entity_element_map.end()) {
-                    qCritical() << "Missing child element";
-                    continue;
-                }
-
-                if (iter->second->is_composite()) {
-                    qCritical() << "Composite child under non-stage composite "
-                                   "is not supported";
-                    continue;
-                }
-
-                auto ret = composite->add_element(iter->second);
-
-                if (ret < 0) { qCritical() << "Add failed"; }
-            }
-        }
-    }
-
-    // Install all elements into the sim engine
-    for (auto const& iter : entity_element_map) {
-        if (m_registry.any_of<ChildOfComponent>(iter.first)) {
-            // Only add top-level elements; composites will add subelements.
-            continue;
-        }
-        auto const& ptr = iter.second;
-        // qDebug() << entt::to_integral(iter.first) << ptr->is_single()
-        //          << ptr->is_composite() << ptr->is_stage();
-        try {
-            ret.add_element(iter.second);
-        } catch (std::exception const& e) {
-            qCritical() << "Unable to export element"
-                        << entt::to_integral(iter.first)
-                        << iter.second->get_name() << e.what();
-
-            return QStringLiteral(
-                "Internal error: an element could not be exported.");
-        }
-    }
-
-    // try {
-    //     SD::write_json_file(ret, "/tmp/soltrace_export.json");
-    // } catch (std::exception const& e) {
-    //     qWarning() << "Failed to write export JSON dump:" << e.what();
-    // }
-
-    std::unordered_map<SD::element_id, entt::entity> entity_rev_map;
-
-    for (auto iter = entity_element_map.begin();
-         iter != entity_element_map.end();
-         ++iter) {
-        entity_rev_map[iter->second->get_id()] = iter->first;
-    }
-
-    auto new_name = QString("%1 v%2")
-                        .arg(this->name())
-                        .arg(QDateTime::currentSecsSinceEpoch());
-
-    auto clone_result =
-        clone_database_with_entity_map(*this, new_name, nullptr);
-
-    for (auto& [element_id, entity] : entity_rev_map) {
-        auto iter = clone_result.old_to_new_map.find(entity);
-        if (iter != clone_result.old_to_new_map.end()) {
-            entity = iter->second;
-        } else {
-            entity = entt::null;
-        }
-    }
-
-    DatabaseExport export_ret;
-    export_ret.data = std::make_shared<SD::SimulationData>(std::move(ret));
-    export_ret.element_map = std::move(entity_rev_map);
-    export_ret.source_database = std::move(clone_result.database);
-
-    return std::make_shared<DatabaseExport>(std::move(export_ret));
 }
 
 QString Database::name() const {
