+++ gui/src/database/database_export.h
@@ -13,8 +13,13 @@ class Database;
 
 /// A packed export of a database to the SolTrace data format.
 struct DatabaseExport {
-    std::shared_ptr<SolTrace::Data::SimulationData>              data;
+    /// SolTrace sim data
+    std::shared_ptr<SolTrace::Data::SimulationData> data;
+
+    /// A map of soltrace element ids to entities
     std::unordered_map<SolTrace::Data::element_id, entt::entity> element_map;
+
+    /// A cloned database from where this data came from
     std::unique_ptr<Database> source_database;
 
     DatabaseExport() = default;
