--- gui/src/database/database_models.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include "database/models/entity_name_model.h"
-#include "database/models/group_models.h"
-#include "database/models/hierarchy_models.h"
-#include "database/models/instance_editor.h"
