+++ gui/src/database/models/instance_editor.cpp
@@ -1,459 +1,14 @@
-#include "database_models.h"
+#include "database/models/instance_editor.h"
 
-#include "components.h"
-#include "conversion.h"
-#include "database/database.h"
-#include "database/database_notification.h"
+#include "database/components.h"
+#include "database/conversion.h"
 #include "utilities/math_utility.h"
 
+#include <glm/geometric.hpp>
+
+#include <unordered_set>
+
 namespace db {
-
-void NameModel::recompute(db::Entity e) {
-    if (!m_host or !e.is_valid() or !m_host->valid(e)) return;
-
-    set_name(m_host->name_of(e));
-}
-
-NameModel::NameModel(QObject* parent) : QObject(parent) { }
-
-void NameModel::reset(Database* database) {
-    if (m_host) {
-        QObject::disconnect(m_host->identity.self(), nullptr, this, nullptr);
-    }
-
-    m_host = database;
-
-    if (!m_host) return;
-
-    connect(m_host->identity.self(),
-            &ComponentAPIBase::changed,
-            this,
-            &NameModel::recompute);
-}
-
-void NameModel::update_name(QString name) {
-    if (!m_host or !m_node.is_valid() or !m_host->valid(m_node)) return;
-
-    if (m_host->name_of(m_node) != m_name) {
-        m_host->identity.patch(
-            m_node, [&](IdentityComponent& ident) { ident.name = m_name; });
-    }
-}
-
-// =============================================================================
-
-void BreadcrumbModel::recompute() {
-    m_path.clear();
-
-    if (!m_host) return;
-
-    if (!m_host->valid(m_node)) return;
-
-    std::unordered_set<entt::entity> seen;
-
-    entt::entity looking_at = m_node;
-
-    while (true) {
-
-        if (!m_host->valid(looking_at)) { break; }
-
-        if (seen.contains(looking_at)) {
-            // cycles??
-            break;
-        }
-
-        seen.insert(looking_at);
-
-        m_path.push_back(looking_at);
-
-        looking_at = m_host->parent_of(looking_at);
-    }
-
-    std::reverse(m_path.begin(), m_path.end());
-
-    QVector<EntityNamePair> ret;
-
-    for (auto e : std::as_const(m_path)) {
-        ret << EntityNamePair { m_host->name_of(e), e };
-    }
-
-    this->store_reset(ret);
-}
-
-BreadcrumbModel::BreadcrumbModel(QObject* parent) : StructModelAdapter(parent) {
-    connect(this,
-            &BreadcrumbModel::node_changed,
-            this,
-            &BreadcrumbModel::recompute);
-}
-
-void BreadcrumbModel::reset(Database* database) {
-    if (m_host) {
-        QObject::disconnect(m_host->identity.self(), nullptr, this, nullptr);
-        QObject::disconnect(m_host->parent.self(), nullptr, this, nullptr);
-    }
-
-    m_host = database;
-    m_node = {};
-    recompute();
-
-    if (database) {
-        connect(database->identity.self(),
-                &ComponentAPIBase::changed,
-                this,
-                [this](entt::entity e) {
-                    if (m_path.contains(Entity(e))) { this->recompute(); }
-                });
-
-        connect(database->parent.self(),
-                &ComponentAPIBase::changed,
-                this,
-                [this](entt::entity e) {
-                    if (m_path.contains(Entity(e))) { this->recompute(); }
-                });
-    }
-}
-
-// =============================================================================
-
-EntityNamePair EntityNamePair::record_for_entity(Database&  db,
-                                                 db::Entity entity) {
-    return EntityNamePair {
-        .name         = db.name_of(entity),
-        .entity       = entity,
-        .has_children = !db.children_of(entity).empty(),
-    };
-}
-
-QVector<EntityNamePair> ChildModel::rebuild_lists() {
-    QVector<EntityNamePair> new_recs;
-    m_reverse.clear();
-
-    if (!m_host) return {};
-
-    if (!m_host->valid(m_node)) return {};
-
-    auto children = m_host->children_of(m_node);
-
-    // Copy here, can revise later
-    new_recs.reserve(children.size());
-
-    for (auto x : children) {
-        new_recs << EntityNamePair::record_for_entity(*m_host, x);
-    }
-
-    for (size_t i = 0; i < new_recs.size(); i++) {
-        m_reverse[new_recs[i].entity] = i;
-    }
-
-    return new_recs;
-}
-
-void ChildModel::recompute() {
-    auto r = rebuild_lists();
-
-    this->store_reset(r);
-}
-
-void ChildModel::record_changed(db::Entity e) {
-    if (!m_host) return;
-
-    if (auto iter = m_reverse.find(e); iter != m_reverse.end()) {
-        this->store_push_update(iter->second,
-                                EntityNamePair::record_for_entity(*m_host, e));
-    }
-}
-
-ChildModel::ChildModel(QObject* parent) : StructModelAdapter(parent) {
-    connect(this, &ChildModel::node_changed, this, &ChildModel::recompute);
-}
-
-void ChildModel::reset(Database* database) {
-    if (m_host) {
-        QObject::disconnect(m_host->children.self(), nullptr, this, nullptr);
-        QObject::disconnect(m_host->identity.self(), nullptr, this, nullptr);
-    }
-
-    m_host = database;
-    m_node = {};
-    recompute();
-
-    if (database) {
-        connect(database->children.self(),
-                &ComponentAPIBase::changed,
-                this,
-                [this](entt::entity e) {
-                    if ((entt::entity)node() == e) recompute();
-                    else
-                        record_changed(db::Entity(e));
-                });
-
-        connect(database->children.self(),
-                &ComponentAPIBase::removed,
-                this,
-                [this](entt::entity e) {
-                    if ((entt::entity)node() == e) recompute();
-                    else
-                        record_changed(db::Entity(e));
-                });
-
-        connect(database->identity.self(),
-                &ComponentAPIBase::changed,
-                this,
-                &ChildModel::record_changed);
-    }
-}
-
-// =============================================================================
-
-QVector<EntityNamePair> MaterialGroupsModel::rebuild_lists() {
-    QVector<EntityNamePair> new_recs;
-    m_reverse.clear();
-
-    if (!m_host) return {};
-
-    auto view = m_host->as_registry().view<MaterialGroupComponent>();
-
-    for (auto const& [e, group] : view.each()) {
-        new_recs.push_back(EntityNamePair::record_for_entity(*m_host, e));
-    }
-
-    for (size_t i = 0; i < new_recs.size(); i++) {
-        m_reverse[new_recs[i].entity] = i;
-    }
-
-    return new_recs;
-}
-
-void MaterialGroupsModel::recompute() {
-    auto r = rebuild_lists();
-
-    this->store_reset(r);
-}
-
-void MaterialGroupsModel::group_changed(db::Entity e) {
-    if (!m_host) return;
-
-    auto iter = m_reverse.find(e);
-
-    if (iter == m_reverse.end()) {
-        if (!m_host->as_registry().all_of<MaterialGroupComponent>(e)) return;
-
-        return recompute();
-    }
-
-    this->store_push_update(iter->second,
-                            EntityNamePair::record_for_entity(*m_host, e));
-}
-void MaterialGroupsModel::group_removed(db::Entity e) {
-    qDebug() << Q_FUNC_INFO << db::Entity(e);
-    recompute();
-}
-
-MaterialGroupsModel::MaterialGroupsModel(QObject* parent)
-    : StructModelAdapter(parent) { }
-
-void MaterialGroupsModel::reset(Database* database) {
-    if (m_host) {
-        QObject::disconnect(m_host->identity.self(), nullptr, this, nullptr);
-        QObject::disconnect(
-            m_host->material_root.self(), nullptr, this, nullptr);
-    }
-    qDebug() << Q_FUNC_INFO << database;
-    m_host = database;
-    recompute();
-
-    if (!database) { return; }
-
-    connect(database->identity.self(),
-            &ComponentAPIBase::changed,
-            this,
-            &MaterialGroupsModel::group_changed);
-
-    connect(database->material_root.self(),
-            &ComponentAPIBase::changed,
-            this,
-            &MaterialGroupsModel::group_changed);
-
-    connect(database->material_root.self(),
-            &ComponentAPIBase::removed,
-            this,
-            &MaterialGroupsModel::group_removed);
-}
-
-QVariant MaterialGroupsModel::get(int index) {
-    auto rec = get_at(index);
-    if (!rec) return {};
-    return QVariant::fromValue(rec->entity);
-}
-
-int MaterialGroupsModel::index_of(db::Entity entity) const {
-    auto iter = m_reverse.find(entity);
-    if (iter == m_reverse.end()) return -1;
-    return iter->second;
-}
-
-// =============================================================================
-
-QVector<EntityNamePair> GeometryGroupsModel::rebuild_lists() {
-    QVector<EntityNamePair> new_recs;
-    m_reverse.clear();
-
-    if (!m_host) return {};
-
-    auto view = m_host->as_registry().view<GeometryGroupComponent>();
-
-    for (auto const& [e, group] : view.each()) {
-        new_recs.push_back(EntityNamePair::record_for_entity(*m_host, e));
-    }
-
-    for (size_t i = 0; i < new_recs.size(); i++) {
-        m_reverse[new_recs[i].entity] = i;
-    }
-
-    return new_recs;
-}
-
-void GeometryGroupsModel::recompute() {
-    auto r = rebuild_lists();
-
-    this->store_reset(r);
-}
-
-void GeometryGroupsModel::group_changed(db::Entity e) {
-    if (!m_host) return;
-
-    auto iter = m_reverse.find(e);
-
-    if (iter == m_reverse.end()) {
-        if (!m_host->as_registry().all_of<GeometryGroupComponent>(e)) return;
-
-        return recompute();
-    }
-
-    this->store_push_update(iter->second,
-                            EntityNamePair::record_for_entity(*m_host, e));
-}
-void GeometryGroupsModel::group_removed(db::Entity e) {
-    recompute();
-}
-
-GeometryGroupsModel::GeometryGroupsModel(QObject* parent)
-    : StructModelAdapter(parent) { }
-
-int GeometryGroupsModel::index_of(db::Entity entity) const {
-    auto iter = m_reverse.find(entity);
-    if (iter == m_reverse.end()) return -1;
-    return iter->second;
-}
-
-void GeometryGroupsModel::reset(Database* database) {
-    if (m_host) {
-        QObject::disconnect(m_host->identity.self(), nullptr, this, nullptr);
-        QObject::disconnect(
-            m_host->geometry_root.self(), nullptr, this, nullptr);
-    }
-    qDebug() << Q_FUNC_INFO << database;
-    m_host = database;
-    recompute();
-
-    if (!database) { return; }
-
-    connect(database->identity.self(),
-            &ComponentAPIBase::changed,
-            this,
-            &GeometryGroupsModel::group_changed);
-
-    connect(database->geometry_root.self(),
-            &ComponentAPIBase::changed,
-            this,
-            &GeometryGroupsModel::group_changed);
-
-    connect(database->geometry_root.self(),
-            &ComponentAPIBase::removed,
-            this,
-            &GeometryGroupsModel::group_removed);
-}
-
-QVariant GeometryGroupsModel::get(int index) {
-    auto rec = get_at(index);
-    if (!rec) return {};
-    return QVariant::fromValue(rec->entity);
-}
-
-// =============================================================================
-
-QVector<EntityNamePair> TagsModel::rebuild_lists() {
-    QVector<EntityNamePair> new_recs;
-    m_reverse.clear();
-
-    if (!m_host) return {};
-
-    auto view = m_host->as_registry().view<TagComponent>();
-
-    for (auto const& e : view.each()) {
-        new_recs.push_back(
-            EntityNamePair::record_for_entity(*m_host, std::get<0>(e)));
-    }
-
-    for (size_t i = 0; i < new_recs.size(); i++) {
-        m_reverse[new_recs[i].entity] = i;
-    }
-
-    return new_recs;
-}
-
-void TagsModel::recompute() {
-    auto r = rebuild_lists();
-
-    this->store_reset(r);
-}
-
-void TagsModel::tag_changed(db::Entity e) {
-    if (!m_host) return;
-
-    auto iter = m_reverse.find(e);
-
-    if (iter == m_reverse.end()) { return recompute(); }
-
-    this->store_push_update(iter->second,
-                            EntityNamePair::record_for_entity(*m_host, e));
-}
-void TagsModel::tag_removed(db::Entity e) {
-    recompute();
-}
-
-TagsModel::TagsModel(QObject* parent) : StructModelAdapter(parent) { }
-
-void TagsModel::reset(Database* database) {
-    if (m_host) {
-        QObject::disconnect(m_host->tag_root.self(), nullptr, this, nullptr);
-        QObject::disconnect(m_host->identity.self(), nullptr, this, nullptr);
-    }
-
-    m_host = database;
-    recompute();
-
-    if (!database) return;
-
-    connect(database->tag_root.self(),
-            &ComponentAPIBase::changed,
-            this,
-            &TagsModel::tag_changed);
-
-    connect(database->tag_root.self(),
-            &ComponentAPIBase::removed,
-            this,
-            &TagsModel::tag_removed);
-
-    connect(database->identity.self(),
-            &ComponentAPIBase::changed,
-            this,
-            &TagsModel::tag_changed);
-}
-
-// =============================================================================
-
 
 #define FIND(MEM)                                                              \
     if (!m_host) return;                                                       \
@@ -490,20 +45,7 @@ void AnInstanceEditor::an_entity_changed(db::Entity e) {
     if (parent() == e) { emit parent_name_changed(); }
 }
 
-AnInstanceEditor::AnInstanceEditor(QObject* parent) : QObject(parent) {
-    /*
-     *     auto lock = m_host.lock();
-
-     if (!lock) { return set_empty(); }
-
-      auto note = get_notifier(*lock);
-
-     if (!note) { return set_empty(); }
-
-      if (!lock->valid(m_entity)) { return set_empty(); }
-
-     */
-}
+AnInstanceEditor::AnInstanceEditor(QObject* parent) : QObject(parent) { }
 
 void AnInstanceEditor::set(db::Entity ent) {
     set_entity(ent);
@@ -634,9 +176,8 @@ QVector3D AnInstanceEditor::position() const {
         }
     }
 
-    return {};
+    return { };
 }
-
 
 void AnInstanceEditor::set_position(const QVector3D& newPosition) {
     if (position() == newPosition) return;
@@ -658,7 +199,7 @@ QVector3D AnInstanceEditor::global_position() const {
         }
     }
 
-    return {};
+    return { };
 }
 
 void AnInstanceEditor::set_global_position(const QVector3D& newPosition) {
@@ -686,7 +227,7 @@ QQuaternion AnInstanceEditor::orientation() const {
             return convert(tf->rotation);
         }
     }
-    return {};
+    return { };
 }
 
 void AnInstanceEditor::set_orientation(const QQuaternion& newOrientation) {
@@ -734,11 +275,10 @@ void AnInstanceEditor::set_hidden(bool newHidden) {
     FIND(invisible);
 
     if (newHidden) {
-        component.set(m_entity, InvisibleComponent {});
+        component.set(m_entity, InvisibleComponent { });
     } else {
         component.remove(m_entity);
     }
-
 
     emit hidden_changed();
 }
@@ -756,7 +296,7 @@ void AnInstanceEditor::set_disabled(bool newDisabled) {
     FIND(disabled);
 
     if (newDisabled) {
-        component.set(m_entity, DisabledComponent {});
+        component.set(m_entity, DisabledComponent { });
     } else {
         component.remove(m_entity);
     }
@@ -785,7 +325,7 @@ db::Entity AnInstanceEditor::material_group() const {
         }
     }
 
-    return {};
+    return { };
 }
 
 void AnInstanceEditor::set_material_group(db::Entity newGroup) {
@@ -807,7 +347,7 @@ db::Entity AnInstanceEditor::geometry_group() const {
         }
     }
 
-    return {};
+    return { };
 }
 
 void AnInstanceEditor::set_geometry_group(db::Entity newGroup) {
@@ -835,7 +375,7 @@ QString AnInstanceEditor::current_material_name() const {
         return m_host->name_of(current_material());
     }
 
-    return {};
+    return { };
 }
 
 Entity AnInstanceEditor::current_geometry() const {
@@ -851,7 +391,7 @@ QString AnInstanceEditor::current_geometry_name() const {
         return m_host->name_of(current_geometry());
     }
 
-    return {};
+    return { };
 }
 
 db::Entity AnInstanceEditor::parent() const {
@@ -859,7 +399,7 @@ db::Entity AnInstanceEditor::parent() const {
         if (auto tf = m_host->parent.get(m_entity); tf) { return tf->parent; }
     }
 
-    return {};
+    return { };
 }
 
 void AnInstanceEditor::set_parent(db::Entity newParent) {
@@ -882,8 +422,9 @@ void AnInstanceEditor::set_parent(db::Entity newParent) {
 
             if (ptr->parent == m_entity) {
                 // cycle
-                emit notify(ANotification::error(
-                    "That parent would create a loop in the element hierarchy."));
+                emit notify(
+                    ANotification::error("That parent would create a loop in "
+                                         "the element hierarchy."));
                 return;
             }
 
@@ -903,7 +444,7 @@ QString AnInstanceEditor::parent_name() const {
         return m_host->name_of(parent());
     }
 
-    return {};
+    return { };
 }
 
 QVector<db::Entity> AnInstanceEditor::tags() const {
@@ -913,7 +454,7 @@ QVector<db::Entity> AnInstanceEditor::tags() const {
         }
     }
 
-    return {};
+    return { };
 }
 
 void AnInstanceEditor::set_tags(QVector<db::Entity> const& newTags) {
@@ -948,7 +489,7 @@ QString AnInstanceEditor::entity_name() const {
         if (auto tf = m_host->identity.get(m_entity); tf) { return tf->name; }
     }
 
-    return {};
+    return { };
 }
 
 void AnInstanceEditor::set_entity_name(const QString& newEntity_name) {
