+++ gui/src/database/models/instance_sort_filter.cpp
@@ -1,175 +1,9 @@
-#include "rootelementsmodel.h"
+#include "database/models/instance_sort_filter.h"
 
 #include "database/components.h"
 #include "database/database_notification.h"
 
 namespace db {
-
-QVector<EntityNamePair> RootElementsModel::rebuild_lists() {
-    QVector<EntityNamePair> new_recs;
-    m_reverse.clear();
-
-    if (!m_host) return { };
-
-    auto view = m_host->as_registry().view<ElementComponent>(
-        entt::exclude<ChildOfComponent>);
-
-    for (auto [entity] : view.each()) {
-        new_recs.push_back(EntityNamePair::record_for_entity(*m_host, entity));
-    }
-
-    for (qsizetype i = 0; i < new_recs.size(); ++i) {
-        m_reverse[new_recs[i].entity] = static_cast<int>(i);
-    }
-
-    return new_recs;
-}
-
-void RootElementsModel::recompute() {
-    store_reset(rebuild_lists());
-}
-
-void RootElementsModel::record_changed(entt::entity entity) {
-    if (!m_host) return;
-
-    if (auto iter = m_reverse.find(entity); iter != m_reverse.end()) {
-        store_push_update(iter->second,
-                          EntityNamePair::record_for_entity(*m_host, entity));
-    }
-}
-
-
-RootElementsModel::RootElementsModel(QObject* parent)
-    : StructModelAdapter(parent) { }
-
-void RootElementsModel::reset(Database* database) {
-    if (m_host) {
-        disconnect(m_host->identity.self(), nullptr, this, nullptr);
-        disconnect(m_host->parent.self(), nullptr, this, nullptr);
-        disconnect(m_host->element_tag.self(), nullptr, this, nullptr);
-        disconnect(m_host->children.self(), nullptr, this, nullptr);
-    }
-
-    m_host = database;
-    recompute();
-
-    if (!database) return;
-
-    connect(database->identity.self(),
-            &ComponentAPIBase::changed,
-            this,
-            &RootElementsModel::record_changed);
-
-    connect(database->children.self(),
-            &ComponentAPIBase::changed,
-            this,
-            &RootElementsModel::record_changed);
-
-    connect(database->children.self(),
-            &ComponentAPIBase::removed,
-            this,
-            &RootElementsModel::record_changed);
-
-    // Root membership is determined by the presence of ChildOfComponent.
-    connect(database->parent.self(),
-            &ComponentAPIBase::changed,
-            this,
-            &RootElementsModel::recompute);
-
-    connect(database->parent.self(),
-            &ComponentAPIBase::removed,
-            this,
-            &RootElementsModel::recompute);
-
-    connect(database->element_tag.self(),
-            &ComponentAPIBase::changed,
-            this,
-            &RootElementsModel::recompute);
-
-    connect(database->element_tag.self(),
-            &ComponentAPIBase::removed,
-            this,
-            &RootElementsModel::recompute);
-}
-
-// =============================================================================
-
-
-QVector<EntityNamePair> AllElementsModel::rebuild_lists() {
-    QVector<EntityNamePair> new_recs;
-    m_reverse.clear();
-
-    if (!m_host) return { };
-
-    auto view = m_host->as_registry().view<ElementComponent>();
-
-    for (auto [entity] : view.each()) {
-        new_recs.push_back(EntityNamePair::record_for_entity(*m_host, entity));
-    }
-
-    for (qsizetype i = 0; i < new_recs.size(); ++i) {
-        m_reverse[new_recs[i].entity] = static_cast<int>(i);
-    }
-
-    return new_recs;
-}
-
-void AllElementsModel::recompute() {
-    store_reset(rebuild_lists());
-}
-
-void AllElementsModel::record_changed(entt::entity entity) {
-    if (!m_host) return;
-
-    if (auto iter = m_reverse.find(entity); iter != m_reverse.end()) {
-        store_push_update(iter->second,
-                          EntityNamePair::record_for_entity(*m_host, entity));
-    }
-}
-
-
-AllElementsModel::AllElementsModel(QObject* parent)
-    : StructModelAdapter(parent) { }
-
-void AllElementsModel::reset(Database* database) {
-    if (m_host) {
-        disconnect(m_host->identity.self(), nullptr, this, nullptr);
-        disconnect(m_host->element_tag.self(), nullptr, this, nullptr);
-        disconnect(m_host->children.self(), nullptr, this, nullptr);
-    }
-
-    m_host = database;
-    recompute();
-
-    if (!database) return;
-
-    connect(database->identity.self(),
-            &ComponentAPIBase::changed,
-            this,
-            &AllElementsModel::record_changed);
-
-    connect(database->children.self(),
-            &ComponentAPIBase::changed,
-            this,
-            &AllElementsModel::record_changed);
-
-    connect(database->children.self(),
-            &ComponentAPIBase::removed,
-            this,
-            &AllElementsModel::record_changed);
-
-    connect(database->element_tag.self(),
-            &ComponentAPIBase::changed,
-            this,
-            &AllElementsModel::recompute);
-
-    connect(database->element_tag.self(),
-            &ComponentAPIBase::removed,
-            this,
-            &AllElementsModel::recompute);
-}
-
-// =============================================================================
 
 void InstanceSortFilter::recompute_has_filter() {
     bool do_name = !m_name_filter.isEmpty();
@@ -178,7 +12,6 @@ void InstanceSortFilter::recompute_has_filter() {
 
     set_has_filter(do_name or do_mat or do_geo);
 }
-
 
 void InstanceSortFilter::update_material_name(Entity entity) {
     if (!m_host) return;
@@ -199,7 +32,6 @@ void InstanceSortFilter::update_geometry_name(Entity entity) {
 InstanceSortFilter::InstanceSortFilter(QObject* parent)
     : QSortFilterProxyModel(parent) {
 
-    // Connect to recompute has filter
     connect(this,
             &InstanceSortFilter::geometry_filter_changed,
             this,
@@ -215,8 +47,6 @@ InstanceSortFilter::InstanceSortFilter(QObject* parent)
             this,
             &InstanceSortFilter::recompute_has_filter);
 
-    // Update names...
-
     connect(this, &InstanceSortFilter::geometry_filter_changed, this, [this]() {
         update_geometry_name(geometry_filter());
     });
@@ -224,8 +54,6 @@ InstanceSortFilter::InstanceSortFilter(QObject* parent)
     connect(this, &InstanceSortFilter::material_filter_changed, this, [this]() {
         update_material_name(material_filter());
     });
-
-    // Invalidate filter on change...
 
     connect(this,
             &InstanceSortFilter::geometry_filter_changed,
@@ -242,7 +70,6 @@ InstanceSortFilter::InstanceSortFilter(QObject* parent)
             this,
             &InstanceSortFilter::invalidate);
 }
-
 
 void InstanceSortFilter::reset(Database* database) {
     if (m_host) {
@@ -273,7 +100,6 @@ void InstanceSortFilter::reset(Database* database) {
             &ComponentAPIBase::changed,
             this,
             &InstanceSortFilter::update_material_name);
-
 
     connect(database->geometry_group_membership.self(),
             &ComponentAPIBase::changed,
@@ -312,14 +138,13 @@ void InstanceSortFilter::clear_all_filters() {
     set_name_filter({ });
 }
 
-bool db::InstanceSortFilter::filterAcceptsRow(
+bool InstanceSortFilter::filterAcceptsRow(
     int                source_row,
     QModelIndex const& source_parent) const {
 
     if (!sourceModel()) return true;
     if (!m_host) return true;
 
-    // TODO: clean this up
     bool do_name = !m_name_filter.isEmpty();
     bool do_mat  = m_material_filter.is_valid();
     bool do_geo  = m_geometry_filter.is_valid();
@@ -358,12 +183,6 @@ bool db::InstanceSortFilter::filterAcceptsRow(
     }
 
     return true;
-}
-
-QVariant RootElementsModel::get(int index) {
-    auto rec = get_at(index);
-    if (!rec) return { };
-    return QVariant::fromValue(rec->entity);
 }
 
 } // namespace db
