+++ gui/src/database/models/instance_sort_filter.h
@@ -1,61 +1,10 @@
 #pragma once
 
-#include "database/database_models.h"
+#include "database/models/entity_name_model.h"
 
 #include <QSortFilterProxyModel>
 
 namespace db {
-
-/// A model providing all element entities that do not currently have a parent.
-class RootElementsModel : public StructModelAdapter<EntityNamePair> {
-    Q_OBJECT
-
-    QPointer<Database> m_host;
-
-    std::unordered_map<entt::entity, int> m_reverse;
-
-    QVector<EntityNamePair> rebuild_lists();
-
-private slots:
-    void recompute();
-    void record_changed(entt::entity);
-
-public:
-    explicit RootElementsModel(QObject* parent = nullptr);
-    ~RootElementsModel() override = default;
-
-    void reset(Database* database);
-
-public slots:
-    QVariant get(int index);
-};
-
-// =============================================================================
-
-/// A model providing all element entities in a database.
-class AllElementsModel : public StructModelAdapter<EntityNamePair> {
-    Q_OBJECT
-
-    QPointer<Database> m_host;
-
-    std::unordered_map<entt::entity, int> m_reverse;
-
-    QVector<EntityNamePair> rebuild_lists();
-
-private slots:
-    void recompute();
-    void record_changed(entt::entity);
-
-public:
-    explicit AllElementsModel(QObject* parent = nullptr);
-    ~AllElementsModel() override = default;
-
-    /// Observe a database and rebuild the element list.
-    void reset(Database* database);
-};
-
-
-// =============================================================================
 
 /// Filter proxy for element models by material, geometry, and name.
 ///
@@ -98,7 +47,6 @@ public slots:
     /// Clear all active filters.
     void clear_all_filters();
 
-    // QSortFilterProxyModel interface
 protected:
     bool filterAcceptsRow(int                source_row,
                           QModelIndex const& source_parent) const override;
