+++ gui/src/database/models/world_geometry_model.cpp
@@ -1,4 +1,4 @@
-#include "worldgeometrymodel.h"
+#include "database/models/world_geometry_model.h"
 #include "utilities/math_utility.h"
 
 #include <algorithm>
