+++ gui/src/database/surface.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <array>
 #include <cstdint>
 #include <optional>
 
@@ -13,9 +12,9 @@
 
 namespace SD = SolTrace::Data;
 
-
 namespace db {
 
+/// Surface generation tunables. Default is 'ok'.
 struct SurfaceGenerationOptions {
     glm::uvec2 height_field_resolution       = { 24, 24 };
     uint32_t   radial_subdivisions           = 24;
@@ -27,10 +26,11 @@ struct SurfaceGenerationOptions {
     double     thickness                     = 0.01;
 
     // Fidelity goes from 1 to 10
-    static SurfaceGenerationOptions from_resolution_and_thickness(unsigned,
-                                                                  float);
+    static SurfaceGenerationOptions
+    from_resolution_and_thickness(unsigned fidelity, float thickness);
 };
 
+/// Compute a mesh from a surface and an aperture.
 std::optional<Mesh>
 generate_surface(SD::surface_ptr const&          surface,
                  SD::aperture_ptr const&         aperture,
